restore date_bug_fix properties_bug_fix restore_watch_contract add_pop_up_functionality

+++ app/src/main/java/com/pixelplex/qtum/utils/DateCalculator.java
@@ -56,7 +56,7 @@ public class DateCalculator {
         return formatter.format(date);
     }
 
-    public static String getDateInFormat(Integer date){
+    public static String getDateInFormat(Long date){
         SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
         return formatter.format(new Date(date));
     }
